Use hwsc-user-svc-psql image for test
