+++ docker-compose.yml
@@ -2,14 +2,13 @@ version: '3'
 
 services:
   postgres:
-    image: "postgres:alpine"
+    image: "hwsc/hwsc-user-svc-psql:test-int"
+    container_name: "test-postgres"
     ports:
       - "5432:5432"
-    environment:
-      - POSTGRES_PASSWORD=$hosts_postgres_password
-      - POSTGRES_DB=$hosts_postgres_db
   user-svc:
     image: "hwsc/hwsc-user-svc:int"
+    container_name: "test-user-svc"
     ports:
       - "50052:50052"
     environment:
